Refactored the balances pallet
