+++ runtime/src/configs/mod.rs
@@ -1,1 +1,2 @@
 pub mod timestamp;
+pub mod balances;
